+++ lib/data/definition/app_mode.dart
