+++ lib/data/usecase/message_use_case.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: prefer-match-file-name
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:isekai/data/definitions/app_mode.dart';
+import 'package:isekai/data/definition/app_mode.dart';
 import 'package:isekai/data/service/database_service.dart';
 import 'package:isekai/data/usecase/session_use_case.dart';
